Updated to comply with manifest_version 2.
Replaced background.html with background.js, tweaked the manifest, this
means this extension will only work in chrome 18 or newer now, but
http://developer.chrome.com/extensions/manifestVersion.html states that
Manifest version 1 items are going to start being heavily phased out of
the web store in the coming months, so it's the best choice to make.
Most people should be on a version newer than 18 anyway.